multiple cars can be spawned from launch file
