+++ src/race/scripts/objects.json.template
